requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paytmchecksum==1.7.0
pycryptodome==3.9.8
python-dotenv==0.14.0
Werkzeug==1.0.1
